feat: Update user last session on sign in
The code changes in `auth.service.ts` update the `signIn` method to set the `lastSession` property of the user to the current date and time when they sign in. This change ensures that the user's last session is accurately tracked and can be used for auditing or session management purposes.