feat: Add current user to recipe creation
The code changes in `recipes.controller.ts` and `recipes.service.ts` modify the `create` method to include the current user information when creating a recipe. This change is achieved by adding the `@CurrentUser()` decorator to the `create` method in the controller and passing the `user` object to the `create` method in the service. By associating the author of the recipe with the current user, we ensure that the recipe is correctly attributed to the user who created it.