feat: Update PDF generation to include user information
The code changes in `supplies.controller.ts` and `users.controller.ts` modify the `generatePdf` method to include the current user information when generating a PDF. This change is achieved by adding the `@CurrentUser()` decorator to the `generatePdf` method in both controllers and passing the `user` object to the `generatePdf` method. By including the user information in the PDF, we ensure that the generated document is personalized and attributed to the correct user.