feat: Add generatePdf endpoint to ProvidersController
The code changes in `providers.controller.ts` add a new endpoint `generatePdf` to the `ProvidersController`. This endpoint generates a PDF file containing a list of providers and their information. The PDF is generated using the `generatePdf` function from the `helpers/handlebars` module. The endpoint sets the necessary headers for the response to be treated as a PDF file attachment. The PDF generation is handled by the `PdfService` class. This addition allows users to download a PDF file with a list of providers.