feat: Add require_maintenance flag to machines
The code changes in `machines.service.ts` introduce a new flag `require_maintenance` to the `Machine` model. This flag is calculated based on the difference in days between the last maintenance date or purchase date and the current date. If the difference is greater than the desired maintenance interval, the flag is set to `true`, indicating that maintenance is required. This addition allows for better tracking and management of machine maintenance.