feat: Add generatePdf endpoint to OrdersController
The code changes in `orders.controller.ts` add a new endpoint `generatePdf` to the `OrdersController`. This endpoint generates a PDF file containing a list of orders and their information. The PDF is generated using the `generatePdf` function from the `helpers/handlebars` module. The endpoint sets the necessary headers for the response to be treated as a PDF file attachment. The PDF generation is handled by the `PdfService` class. This addition allows users to download a PDF file with a list of orders.